Refactor request-password
